move some methods from convolution namespace to more sensible namespaces
to tools, and to new ir namespace for ir manipulation (which convolution shouldnt care about because IRs are input for it)
